_codegen/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
docker==6.1.3
frozenlist==1.4.0
idna==3.4
loguru==0.7.0
multidict==6.0.4
openai==0.27.8
packaging==23.1
promptlayer==0.1.98
pydantic==1.10.12
python-dotenv==1.0.0
requests==2.31.0
tenacity==8.2.3
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
websocket-client==1.6.1
yarl==1.9.2
